test: move shared code in to conftest
